1.11.1.0

+++ app/code/core/Enterprise/PageCache/Model/Observer.php
@@ -410,9 +410,11 @@ class Enterprise_PageCache_Model_Observer
         }
         $this->_getCookie()->updateCustomerCookies();
 
-        $this->_getCookie()->delete(Enterprise_PageCache_Model_Cookie::COOKIE_RECENTLY_COMPARED);
-        $this->_getCookie()->delete(Enterprise_PageCache_Model_Cookie::COOKIE_COMPARE_LIST);
-        Enterprise_PageCache_Model_Cookie::registerViewedProducts(array(), 0, false);
+        if (!$this->_getCookie()->get(Enterprise_PageCache_Model_Cookie::COOKIE_CUSTOMER)) {
+            $this->_getCookie()->delete(Enterprise_PageCache_Model_Cookie::COOKIE_RECENTLY_COMPARED);
+            $this->_getCookie()->delete(Enterprise_PageCache_Model_Cookie::COOKIE_COMPARE_LIST);
+            Enterprise_PageCache_Model_Cookie::registerViewedProducts(array(), 0, false);
+        }
 
         return $this;
     }
@@ -534,5 +536,46 @@ class Enterprise_PageCache_Model_Observer
             }
             Mage::getSingleton('core/cookie')->delete($varName);
         }
+    }
+
+    /**
+     * Update info about product on product page
+     *
+     * @param Varien_Event_Observer $observer
+     * @return Enterprise_PageCache_Model_Observer
+     */
+    public function updateProductInfo(Varien_Event_Observer $observer)
+    {
+        $paramsObject = $observer->getEvent()->getParams();
+        if ($paramsObject instanceof Varien_Object) {
+            if (array_key_exists(Enterprise_PageCache_Model_Cookie::COOKIE_CATEGORY_ID, $_COOKIE)) {
+                $paramsObject->setCategoryId($_COOKIE[Enterprise_PageCache_Model_Cookie::COOKIE_CATEGORY_ID]);
+            }
+        }
+        return $this;
+    }
+
+    /**
+     * Check cross-domain session messages
+     *
+     * @param Varien_Event_Observer $observer
+     * @return Enterprise_PageCache_Model_Observer
+     */
+    public function checkMessages(Varien_Event_Observer $observer)
+    {
+        $transport = $observer->getEvent()->getTransport();
+        if (!$transport || !$transport->getUrl()) {
+            return $this;
+        }
+        $url = $transport->getUrl();
+        $httpHost = Mage::app()->getFrontController()->getRequest()->getHttpHost();
+        $urlHost = parse_url($url, PHP_URL_HOST);
+        if ($httpHost != $urlHost && Mage::getSingleton('core/session')->getMessages()->count() > 0) {
+            $transport->setUrl(Mage::helper('core/url')->addRequestParam(
+                $url,
+                array(Enterprise_PageCache_Model_Cache::REQUEST_MESSAGE_GET_PARAM => null)
+            ));
+        }
+        return $this;
     }
 }
